requirements.txt
pycryptodomex==3.19.0
pyOpenSSL==23.3.0
pyOpenSSL==23.3.0
